merge "119.iso8601 cycling" to master, for cylc-6

+++ tests/vacation/00-sigusr1.t
@@ -15,28 +15,45 @@
 #C: You should have received a copy of the GNU General Public License
 #C: along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #-------------------------------------------------------------------------------
-#C: Test handle of SIGUSR1. (Handle a mock job vacation.)
-#C: Obviously, job vacation does not happen with background job, and the job
-#C: will no longer be poll-able after the kill.
+# Test handle of SIGUSR1. (Handle a mock job vacation.)
+# Obviously, job vacation does not happen with background job, and the job
+# will no longer be poll-able after the kill.
 . $(dirname $0)/test_header
 #-------------------------------------------------------------------------------
-set_test_number 6
+set_test_number 8
 install_suite $TEST_NAME_BASE $TEST_NAME_BASE
 #-------------------------------------------------------------------------------
 TEST_NAME=$TEST_NAME_BASE-validate
 run_ok $TEST_NAME cylc validate $SUITE_NAME
 #-------------------------------------------------------------------------------
 TEST_NAME=$TEST_NAME_BASE-run
-run_ok $TEST_NAME cylc run --reference-test $SUITE_NAME
+suite_run_ok $TEST_NAME cylc run --reference-test $SUITE_NAME
 #-------------------------------------------------------------------------------
 SUITE_RUN_DIR=$(cylc get-global-config --print-run-dir)/$SUITE_NAME
+
 # Make sure t1.1.1's status file is in place
 T1_STATUS_FILE=$SUITE_RUN_DIR/log/job/t1.1.1.status
+
+TEST_NAME=$TEST_NAME_BASE-find-status-file
 TIMEOUT=$(($(date +%s) + 120))
-while [[ ! -f $T1_STATUS_FILE ]] && (($TIMEOUT > $(date +%s))); do
+while [[ ! -f $T1_STATUS_FILE ]]; do
     sleep 1
+    if (($(date +%s) > $TIMEOUT)); then
+        fail $TEST_NAME
+        exit 1
+    fi
 done
+ok $TEST_NAME
+
+# Read the process id from the file.
+TEST_NAME=$TEST_NAME_BASE-get-pid-from-status-file
 T1_PID=$(awk -F= '$1=="CYLC_JOB_PID" {print $2}' $T1_STATUS_FILE)
+if [[ -z $T1_PID ]]; then
+    fail $TEST_NAME
+    exit 1
+fi
+ok $TEST_NAME
+
 # Kill the job and see what happens
 kill -s USR1 $T1_PID
 while ps $T1_PID 1>/dev/null 2>&1; do
